Import the types used in repositories from packages/shared and organize them in types/inputs.ts.
