+++ src/types/inputs.ts
@@ -1,0 +1,6 @@
+export { upsertAssignShfitInputType } from "@shared/shift/assign/validations/put";
+export { CreatePaymentInput } from "@shared/payment/types/post";
+export { UpsertShiftRequetType } from "@shared/shift/request/validations/put";
+export { upsertSubmittedShiftInputType } from "@shared/shift/submit/validations/put";
+export { UpsertUserInput } from "@shared/auth/types/register-owner";
+export { updateUserProlfileType } from "@shared/user/validations/put";
